Move reset() function up to RedwoodView
